requirements.txt
PyTelegramBotAPI
pytesseract
python-magic
Pillow
